requirements.txt
requests
python-dotenv
kafka-python
pyspark
numpy
pandas
statsmodels
pytest
typing_extensions
